+++ backend/models/follow.js
@@ -18,5 +18,20 @@ module.exports = (sequelize, DataTypes) => {
       modelName: "Follow",
     }
   );
+
+  Follow.getFollowingIdLIst = async (user_id) => {
+    const followingIdList = [];
+    const following = await Follow.findAll({
+      where: { follower_id: user_id, flag: true },
+      attributes: ["following_id"],
+    });
+
+    if (following.length > 0) {
+      following.map((m) => {
+        followingIdList.push(m.following_id);
+      });
+    }
+    return followingIdList;
+  };
   return Follow;
 };
